% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandiff.R
\name{ci_width_md_equal}
\alias{ci_width_md_equal}
\title{Calculates the probability, width or sample size given the two others
for confidence interval of a mean difference.}
\usage{
ci_width_md_equal(
  sigma,
  width = NULL,
  n1 = NULL,
  prob = NULL,
  nratio = 1,
  conf.level = 0.95,
  plot = FALSE,
  plot_order = c("n1", "width", "prob", "sigma", "nratio"),
  connect = TRUE
)
}
\arguments{
\item{sigma}{Common standard deviation.}

\item{width}{Width of confidence interval.}

\item{n1}{Number of observations in group 1.}

\item{prob}{Probability of obtaining given width.}

\item{nratio}{Ratio of group sizes (n2/n1).}

\item{conf.level}{Confidence level of the interval.}

\item{plot}{\code{TRUE} returns a \code{ggplot2} plot. \code{FALSE} returns a data.frame.}

\item{plot_order}{Names of the variables to display in the plot.
\itemize{
\item A vector with names of up to four arguments (sigma, width, n, prob, conf.level)
\item The order of the arguments determines the appearance of the plot:
\itemize{
\item 1: y-axis, 2: x-axis, 3: color, 4: facet.
}
}}

\item{connect}{Connects the observations in the plot.}
}
\value{
A \code{ggplot2} object or data.frame (if \code{plot = FALSE}).
}
\description{
One can provide several values of arguments, and the function returns
the results as a data.frame or a \code{ggplot2} plot.
}
\details{
This assumes the observations in each group are i.i.d. normally distributed
with a common standard deviation sigma which is treated as unknown to us
and must be estimated.
}
\examples{
# Calculate the sample sizes needed to obtain a 95\% confidence interval for
# a mean difference of width 1, 3 and 5 with a probability of either 80\%
# or 90\% and a standard deviation of 10 or 15 in both groups:
ci_width_md_equal(sigma = c(10, 15), width = c(1, 3, 5), prob = c(0.80, 0.90))
# Make a plot of the sample size as a function of width, probability and
# standard deviation:
ci_width_md_equal(sigma = c(10, 15), width = c(1, 3, 5), prob = c(0.80, 0.90), plot = TRUE)
}
