% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{ci_width_mean}
\alias{ci_width_mean}
\title{Calculates the probability, width or sample size given the two others
for confidence interval of a mean.}
\usage{
ci_width_mean(
  sigma,
  width = NULL,
  n = NULL,
  prob = NULL,
  conf.level = 0.95,
  plot = FALSE,
  plot_order = c("n", "width", "prob", "sigma")
)
}
\arguments{
\item{sigma}{Standard deviation.}

\item{width}{Width of confidence interval.}

\item{n}{Number of observations.}

\item{prob}{Probability of obtaining given width.}

\item{conf.level}{Confidence level of the interval.}

\item{plot}{\code{TRUE} returns a \code{ggplot2} plot. \code{FALSE} returns a data.frame.}

\item{plot_order}{Names of the variables to display in the plot.
\itemize{
\item A vector with names of up to four arguments (sigma, width, n, prob, conf.level)
\item The order of the arguments determines the appearance of the plot:
\itemize{
\item 1: y-axis, 2: x-axis, 3: color, 4: facet.
}
}}
}
\value{
A \code{ggplot2} object or data frame (if \code{plot = FALSE}).
}
\description{
One can provide several values of arguments, and the function returns
the results as a data.frame or a \code{ggplot2} plot.
}
\details{
This assumes the observations are normally distributed with standard
deviation sigma which is treated as unknown to us and must be estimated.
}
\examples{
# Calculate the sample sizes needed to obtain a 95\% confidence interval
# of width 5, 6, 7, 8 and 10 with a probability of either 80\% or 90\% a
# standard deviation of 10 or 15:
ci_width_mean(sigma = c(10, 15), width = 5:10, prob = c(0.80, 0.90))

# Make a plot of the sample size as a function of width, probability and
# standard deviation:
ci_width_mean(sigma = c(10, 15), width = 5:10, prob = c(0.80, 0.90), plot = TRUE)
}
